fixed lint issue

+++ common/txmgr/confirmer.go
@@ -1074,7 +1074,7 @@ func (ec *Confirmer[CHAIN_ID, HEAD, ADDR, TX_HASH, BLOCK_HASH, R, SEQ, FEE]) han
 // re-org'd out and will be rebroadcast.
 func (ec *Confirmer[CHAIN_ID, HEAD, ADDR, TX_HASH, BLOCK_HASH, R, SEQ, FEE]) EnsureConfirmedTransactionsInLongestChain(ctx context.Context, head types.Head[BLOCK_HASH], latestFinalizedHeadNumber int64) error {
 	earliestInChain := head.EarliestHeadInChain()
-	chainLength := uint32(head.BlockNumber() - earliestInChain.BlockNumber())
+	chainLength := head.BlockNumber() - earliestInChain.BlockNumber()
 	logArgs := []interface{}{
 		"chainLength", chainLength, "latestFinalizedHead number", latestFinalizedHeadNumber,
 	}
@@ -1085,7 +1085,7 @@ func (ec *Confirmer[CHAIN_ID, HEAD, ADDR, TX_HASH, BLOCK_HASH, R, SEQ, FEE]) Ens
 		return errors.New(errMsg)
 	}
 
-	calculatedFinalityDepth := uint32(head.BlockNumber() - latestFinalizedHeadNumber)
+	calculatedFinalityDepth := head.BlockNumber() - latestFinalizedHeadNumber
 	if chainLength < calculatedFinalityDepth {
 		if ec.nConsecutiveBlocksChainTooShort > logAfterNConsecutiveBlocksChainTooShort {
 			warnMsg := "Chain length supplied for re-org detection was shorter than the depth from the latest head to the finalized head. Re-org protection is not working properly. This could indicate a problem with the remote RPC endpoint, a compatibility issue with a particular blockchain, a bug with this particular blockchain, heads table being truncated too early, remote node out of sync, or something else. If this happens a lot please raise a bug with the Chainlink team including a log output sample and details of the chain and RPC endpoint you are using."
